# Optimize Access.filter to eliminate intermediate list creation